+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/laf/DefaultLookAndFeelEnabler.java
@@ -19,6 +19,7 @@
 package is.codion.swing.common.ui.laf;
 
 import is.codion.swing.common.ui.Utilities;
+import is.codion.swing.common.ui.font.FontSize;
 
 import javax.swing.LookAndFeel;
 import javax.swing.UIManager;
@@ -113,6 +114,7 @@ final class DefaultLookAndFeelEnabler implements LookAndFeelEnabler {
 		public void accept(LookAndFeelInfo lookAndFeelInfo) {
 			try {
 				UIManager.setLookAndFeel(requireNonNull(lookAndFeelInfo).getClassName());
+				FontSize.updateFontSize();
 				Utilities.updateComponentTreeForAllWindows();
 			}
 			catch (Exception e) {