+++ swing/common-ui/src/main/java/module-info.java
@@ -66,6 +66,7 @@ module is.codion.swing.common.ui {
 	exports is.codion.swing.common.ui.control;
 	exports is.codion.swing.common.ui.cursor;
 	exports is.codion.swing.common.ui.dialog;
+	exports is.codion.swing.common.ui.font;
 	exports is.codion.swing.common.ui.frame;
 	exports is.codion.swing.common.ui.icon;
 	exports is.codion.swing.common.ui.key;